-fixed bug with not found song

+++ downloader.py
@@ -34,26 +34,27 @@ def download_songs(artist, album, tracks, file_path, status_handler):
         song_name = folder_directory + '/' + track_name + '.mp3'
 
         # Download track
-        urllib.request.urlretrieve(track['trackUrl'], song_name)
+        if 'trackUrl' in track:
+            urllib.request.urlretrieve(track['trackUrl'], song_name)
 
-        # Metadata correction
-        metadata = MP3(song_name)
-        print(metadata.tags)
-        metadata.delete()
-        # Artist
-        metadata['composer'] = artist['artistName']
-        metadata['performer'] = artist['artistName']
-        metadata['artist'] = artist['artistName']
+            # Metadata correction
+            metadata = MP3(song_name)
+            print(metadata.tags)
+            metadata.delete()
+            # Artist
+            metadata['composer'] = artist['artistName']
+            metadata['performer'] = artist['artistName']
+            metadata['artist'] = artist['artistName']
 
-        metadata['album'] = track['collectionName']
-        metadata['title'] = track['trackName']
-        metadata['date'] = track['releaseDate']
-        metadata['genre'] = track['primaryGenreName']
-        metadata['tracknumber'] = str(track['trackNumber'])
-        metadata['discnumber'] = str(track['discNumber'])
+            metadata['album'] = track['collectionName']
+            metadata['title'] = track['trackName']
+            metadata['date'] = track['releaseDate']
+            metadata['genre'] = track['primaryGenreName']
+            metadata['tracknumber'] = str(track['trackNumber'])
+            metadata['discnumber'] = str(track['discNumber'])
 
-        metadata.update()
-        metadata.save()
+            metadata.update()
+            metadata.save()
 
         status_handler('Song uploaded', progress)
         progress += tick
